chore: Refactor SQS handler to a separate class
